backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.3
brotli==1.2.0
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cssselect2==0.8.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
fonttools==4.60.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.0.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
reportlab==4.4.5
requests==2.32.5
rpds-py==0.29.0
rsa==4.9.1
sentry-sdk==2.45.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
gunicorn==21.2.0
